Move layer routing function to abstract class
First draft of separating the routing functions of StreetLayer from the OSM-based network building functions. The idea is that old StreetLayer becomes OSMStreetLayer, and other implementations can extend RoutableStreetLayer to provide a routable network layer built from non-OSM sources.
